+++ dwtools/Categories.cpp
@@ -24,7 +24,7 @@
 #include "Categories.h"
 
 void structCategories :: v_readText (MelderReadText a_text, int /*formatVersion*/) {
-	long l_size = texgeti32 (a_text);
+	integer l_size = texgeti32 (a_text);
 	if (l_size == 0) {
 		(void) 0;
 	} else if (l_size < 0) {
@@ -32,7 +32,7 @@ void structCategories :: v_readText (MelderReadText a_text, int /*formatVersion*
 	} else {
 		our _grow (l_size);
 	}
-	for (long i = 1; i <= l_size; i ++) {
+	for (integer i = 1; i <= l_size; i ++) {
 		autoSimpleString itemi = Thing_new (SimpleString);
 		itemi -> v_readText (a_text, 0);
 		our addItemAtPosition_move (itemi.move(), i);
@@ -41,7 +41,7 @@ void structCategories :: v_readText (MelderReadText a_text, int /*formatVersion*
 
 void structCategories :: v_writeText (MelderFile file) {
 	texputi32 (file, our size, U"size", nullptr, nullptr, nullptr, nullptr, nullptr);
-	for (long i = 1; i <= our size; i ++) {
+	for (integer i = 1; i <= our size; i ++) {
 		SimpleString data = our at [i];
 		texputintro (file, U"item [", Melder_integer (i), U"]:", nullptr, nullptr, nullptr);
 		data -> structSimpleString :: v_writeText (file);
@@ -60,7 +60,7 @@ autoCategories Categories_create () {
 	}
 }
 
-autoCategories Categories_createWithSequentialNumbers (long n) {
+autoCategories Categories_createWithSequentialNumbers (integer n) {
 	try {
 		autoCategories me = Thing_new (Categories);
 		OrderedOfString_initWithSequentialNumbers (me.get(), n);
@@ -80,7 +80,7 @@ autoCategories Categories_selectUniqueItems (Categories me) {
 	}
 }
 
-void Categories_drawItem (Categories me, Graphics g, long position, double xWC, double yWC) {
+void Categories_drawItem (Categories me, Graphics g, integer position, double xWC, double yWC) {
 	if (position < 1 || position > my size) {
 		return;
 	}
@@ -91,7 +91,7 @@ autoCategories OrderedOfString_to_Categories (OrderedOfString me) {
 	try {
 		autoCategories thee = Categories_create();
 
-		for (long i = 1; i <= my size; i ++) {
+		for (integer i = 1; i <= my size; i ++) {
 			autoSimpleString item = Data_copy (my at [i]);
 			thy addItem_move (item.move());
 		}
@@ -101,7 +101,7 @@ autoCategories OrderedOfString_to_Categories (OrderedOfString me) {
 	}
 }
 
-long Categories_getSize (Categories me) {
+integer Categories_getSize (Categories me) {
 	return my size;
 }
 
@@ -110,8 +110,8 @@ autoCategories TableOfReal_to_CategoriesRow (TableOfReal me) {
 	try {
 		autoCategories thee = Categories_create ();
 
-		for (long i = 1; i <= my numberOfRows; i ++) {
-			if (my rowLabels[i]) {
+		for (integer i = 1; i <= my numberOfRows; i ++) {
+			if (my rowLabels [i]) {
 				autoSimpleString s = SimpleString_create (my rowLabels [i]);
 				thy addItem_move (s.move());
 			}
@@ -126,8 +126,8 @@ autoCategories TableOfReal_to_CategoriesColumn (TableOfReal me) {
 	try {
 		autoCategories thee = Categories_create ();
 
-		for (long i = 1; i <= my numberOfColumns; i ++) {
-			if (my columnLabels[i]) {
+		for (integer i = 1; i <= my numberOfColumns; i ++) {
+			if (my columnLabels [i]) {
 				autoSimpleString s = SimpleString_create (my columnLabels [i]);
 				thy addItem_move (s.move());
 			}
